+++ src/@types/image.ts
@@ -6,3 +6,7 @@ declare module "*.svg" {
     const value: string;
     export default value;
 }
+declare module "*.htm" {
+    const value: string;
+    export default value;
+}
