+++ src/jsx/TopBar/BugReportButton.tsx
@@ -1,23 +1,20 @@
 import {Button} from '@blueprintjs/core';
-import { getClient } from "@sentry/react"
 import LadyBug from 'Images/gustavorezende_lady_bug-555px.png';
-import React, {useRef, useEffect} from 'react';
+import {useRef, useEffect} from 'react';
 import {useTranslation} from 'react-i18next'
 import { useMediaQuery } from "react-responsive";
+import * as Sentry from "@sentry/react";
 
 const BugReportButton = () => {
-        const buttonRef = useRef()
+        const buttonRef = useRef<HTMLButtonElement>(null)
         const buttonDimensions = useMediaQuery({ query: '(min-width: 1300px)' }) ? { height: "45", width: "32" } : { height: "34", width: "25" }
         const buttonHeight = useMediaQuery({ query: '(min-width: 1300px)' }) ? '47px' : '36px'
         const { t } = useTranslation()
         useEffect(() => {
-                const client = getClient();
-                if (client) {
-                        const feedback = client.getIntegrationByName("Feedback")
-                        if (buttonRef.current) {
-                                feedback.attachTo(buttonRef.current)
-                        }
-                }        
+                const feedback = Sentry.feedbackIntegration({autoInject:false, isEmailRequired:false})
+                if (buttonRef.current && feedback) {
+                        feedback.attachTo(buttonRef.current)
+                }
         })
         return (
                 <Button className="pt-intent-warning"
